refactor: update CBlockIndex::nChainTx to be uint64_t
There was a comment with this data member: "Change to 64-bit type before 2024 (assuming worst case of 60 byte transactions)."
As of when this PR was created, there were only around 750,000,000 transactions in the best chain.
I went through all the references to this member and made updates to make the data types line up better where there were differences.

+++ src/chain.h
@@ -183,14 +183,13 @@ public:
     unsigned int nTx{0};
 
     //! (memory only) Number of transactions in the chain up to and including this block.
-    //! This value will be non-zero only if and only if transactions for this block and all its parents are available.
-    //! Change to 64-bit type before 2024 (assuming worst case of 60 byte transactions).
+    //! This value will be non-zero if, and only if, transactions for this block and all its parents are available.
     //!
     //! Note: this value is faked during use of a UTXO snapshot because we don't
     //! have the underlying block data available during snapshot load.
     //! @sa AssumeutxoData
     //! @sa ActivateSnapshot
-    unsigned int nChainTx{0};
+    uint64_t nChainTx{0};
 
     //! Verification status of this block. See enum BlockStatus
     //!
